Add fix for proper TS padding for SCTE-35 messages
When padding is required for table based content (i.e. not PES)
packets, the correct behavior is to not use Adaptation padding but
rather to use the alternative stuffing scheme defined in ISO 13818-1
Sec 2.4.4).

I confirmed that this is behavior consistent with SCTE-35 messages
generated by Adtec, Ateme, and Elemental encoders.

This is required for interoperability with Ericsson RX8200 decoders
which do not support SCTE-35 triggers that use Adaptation padding.